requirements.txt
Flask==2.3.2
python-docx==0.8.11
PyPDF2==3.0.0
pypdf==4.0.0
werkzeug==3.0.1
pdfplumber==0.10.3
pillow
rapidfuzz






